etnaviv: add SIN and COS implementations using Taylor series
Add SIN and COS implementations for GPUs which do not have native
SIN and COS instructions using a Taylor series.

This provides an incomplete implementation: we should MOD the input
value to 1..-1 but as we have no MOD instruction support, we omit
this step for the time being.

This fixes the SIN/COS piglit glean tests on GC600.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>