etnaviv: avoid using tile status on MC1.0 GPUs
Kernel sources suggest that there is a requirement that the tile status
buffers on MC1.0 GPUs are located in the linear memory region, and also
that the tile status module addresses are not affected by the GPU
physical address offset registers.

Since we always program an offset address for all relocations, and we
are unable to force allocations out of the GPU, disable use of the tile
status module on MC1.0 GPUs.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>